requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
python-dotenv>=1.0
pdfminer.six>=20231228
python-docx>=1.1
PyPDF2>=3.0
spacy>=3.8
scikit-learn>=1.5
reportlab>=4.2
pytest>=8.3
pytest-django>=4.8
